Feature/public private template projects boards

+++ api/controllers/boardTemplate.controllers.js
@@ -2,9 +2,59 @@ import BoardModel from "../models/Board.model.js";
 import BoardTemplateModel from "../models/BoardTemplate.model.js";
 import TaskModel from "../models/Task.model.js";
 
-export async function getBoardsTemplates(req, res, next) {
+
+
+export async function saveBoardTemplate(req, res, next) {
   try {
-    const boardTemplates = await BoardTemplateModel.find();
+    const authUser = res.locals.user;
+    const { name, boardId, isPrivate } = req.body;
+
+    // Check if the required fields are provided
+    if (!boardId || !name) {
+      return res.status(400).send({
+        success: false,
+        message: "Paramètres manquants",
+      });
+    }
+
+    const board = await BoardModel.findById({ _id: boardId });
+
+    if (!board) {
+      return res.status(404).send({
+        success: false,
+        message: "Impossible de créer le modèle d'un tableau qui n'existe pas",
+      });
+    }
+
+    const newTemplate = new BoardTemplateModel({
+      name: name,
+      author: authUser?._id,
+      boardId: boardId,
+      private: isPrivate,
+    });
+
+    const savedTemplate = await newTemplate.save();
+
+    return res.status(201).send({
+      success: true,
+      message: "Modèle de tableau enregistré avec succès",
+      data: savedTemplate,
+    });
+  } catch (err) {
+    return res.status(500).send({
+      success: false,
+      message:
+        err?.message ||
+        "Une erreur s'est produite lors de l'enregistrement du modèle",
+    });
+  }
+}
+
+export async function getPublicBoardsTemplates(req, res, next) {
+  try {
+    const boardTemplates = await BoardTemplateModel.find({
+      private: false
+    });
 
     if (boardTemplates.length <= 0) {
       return res.status(404).send({
@@ -28,47 +78,33 @@ export async function getBoardsTemplates(req, res, next) {
   }
 }
 
-export async function saveBoardTemplate(req, res, next) {
+export async function getUserPrivateBoardsTemplates(req, res, next) {
   try {
-    const authUser = res.locals.user;
-    const { name, boardId } = req.body;
+    const authUser = res.locals.user
 
-    // Check if the required fields are provided
-    if (!boardId || !name) {
-      return res.status(400).send({
-        success: false,
-        message: "Paramètres manquants",
-      });
-    }
-
-    const board = await BoardModel.findById({ _id: boardId });
-
-    if (!board) {
-      return res.status(404).send({
-        success: false,
-        message: "Impossible de créer le modèle d'un tableau qui n'existe pas",
-      });
-    }
-
-    const newTemplate = new BoardTemplateModel({
-      name: name,
+    const boardTemplates = await BoardTemplateModel.find({
       author: authUser?._id,
-      boardId: boardId,
+      private: true
     });
 
-    const savedTemplate = await newTemplate.save();
+    if (boardTemplates.length <= 0) {
+      return res.status(404).send({
+        success: false,
+        message: "Aucun modèle trouvé",
+      });
+    }
 
-    return res.status(201).send({
+    return res.status(200).send({
       success: true,
-      message: "Modèle de tableau enregistré avec succès",
-      data: savedTemplate,
+      message: "Modèles récupérés avec succès",
+      data: boardTemplates,
     });
   } catch (err) {
     return res.status(500).send({
       success: false,
       message:
         err?.message ||
-        "Une erreur s'est produite lors de l'enregistrement du modèle",
+        "Une erreur s'est produite lors de la récupération des modèles",
     });
   }
 }
@@ -130,16 +166,27 @@ export async function useBoardTemplate(req, res, next) {
 
 export async function deleteBoardTemplate(req, res, next) {
   try {
-    const deletedTemplate = await BoardTemplateModel.findByIdAndDelete({
-      _id: req.params.id,
-    });
+    const authUser = res.locals.user;
 
-    if (!deletedTemplate) {
+    const template = await BoardTemplateModel.findOne({ _id: req.params.id });
+    console.log("template :", template)
+    if (!template) {
       return res.status(404).send({
         success: false,
         message: "Modèle non trouvé",
       });
     }
+
+    if (template?.author?.toString() !== authUser?._id?.toString()) {
+      return res.status(403).send({
+        success: false,
+        message: "Vous n'avez pas les permissions pour supprimer ce modèle",
+      });
+    }
+
+    const deletedTemplate = await BoardTemplateModel.findByIdAndDelete({
+      _id: req.params.id,
+    });
 
     await TaskModel.deleteMany({ boardId: deletedTemplate._id });
